Upgraded tests to use QUnit and made them more in depth.
I wasn't impressed with the lackluster testing and I thought this
project deserved better which is why I decided to go with the
same unit testing framework that jQuery uses. If it's good
enough for them then it should be good enough for us.

Right?

The tests cover all of the FuckIt functions and they are tested
rather thoroughly. Hopefully this works for everyone. I'm sure
it will.

I really can't see anything wrong with the above code.